compare

+++ LNKWallet/waitingforsync.cpp
@@ -119,7 +119,7 @@ int compareVersion( QString a, QString b)
     {
         return 1;
     }
-    else if( aList.at(0) < bList.at(0))
+    else if( aList.at(0).toInt() < bList.at(0).toInt())
     {
         return -1;
     }
@@ -129,7 +129,7 @@ int compareVersion( QString a, QString b)
         {
             return 1;
         }
-        else if( aList.at(1) < bList.at(1))
+        else if( aList.at(1).toInt() < bList.at(1).toInt())
         {
             return -1;
         }
@@ -139,7 +139,7 @@ int compareVersion( QString a, QString b)
             {
                 return 1;
             }
-            else if( aList.at(2) < bList.at(2))
+            else if( aList.at(2).toInt() < bList.at(2).toInt())
             {
                 return -1;
             }
